Collision detected
Detecting collisions, however it's weirdly causing both objects to
continue moving until they come apart. Can also pass through (albeit
reluctantly) if you hold the move key. Cout statements confirm that it
is recognising collisions though. Also, Sam can't figure out why having
the height set (correctly) to 2 in the cPhysicsComponents file is
causing it to float above the plane. If I set it to 1, it sits nicely on
the plane, however the collision will only be detected in the bottom
half of the object. This doesn't really matter at this point, but going
forward the dominoes will tip and contact will need to be made in the
upper region of the second object.